read_hdf: while preprocessing only read each node once
The HDF file format is built in a way that makes reading attributes relatively
slow and requires reading the data as well. For that reason it may be worth our
while to iterate once for both key and storer instead of the previous approach
which tries to first filter the required keys and only then read the required
storers.

Signed-off-by: Nir Izraeli <nirizr@gmail.com>